requirements.txt
# OpenCure Labs — Python Dependencies
# Install: pip install -r requirements.txt
# Versions pinned to known-working set to prevent pip resolver backtracking.
# Update with: pip freeze | grep <package> after testing locally.

# ── Core Framework ───────────────────────────────────────────────────────────
pydantic==2.12.5
psycopg2-binary==2.9.11
requests==2.33.0
pyyaml==6.0.3

# ── NVIDIA NeMo Agent Toolkit ────────────────────────────────────────────────
nvidia-nat==1.5.0

# ── Numerical & Data ─────────────────────────────────────────────────────────
numpy==2.4.3
pandas==2.3.3
scikit-learn==1.8.0
pyarrow==23.0.1

# ── Scientific Stack (neoantigen pipeline) ───────────────────────────────────
pysam==0.23.3
biopython==1.86
# pyensembl and mhcflurry have transitive deps that conflict with pyarrow 23.x
# (gtfparse pins pyarrow<14.1). Install separately: see .github/workflows/ci.yml
pyensembl==2.3.13
mhcflurry==2.1.5
rdkit==2025.9.6

# ── LangGraph / LangChain (coordinator agent) ───────────────────────────────
# langchain-core >= 1.2.31 required for SSRF-safe transport used by
# langchain-openai 1.1.14 (GHSA-r7w7-9xr2-qq2r).
langchain-core==1.2.31
langchain-openai==1.1.14
# langsmith 0.7.31 closes streaming-output redaction bypass (GHSA-rr7j-v2q5-chgv).
langsmith==0.7.32
# authlib 1.6.11 adds CSRF protection on the cache feature (GHSA-jj8c-mmj3-mmgv).
authlib==1.6.11
langgraph==1.1.3
langgraph-prebuilt==1.0.8  # pin: 1.0.9 requires ServerInfo not in langgraph 1.1.3

# ── Reviewer Dependencies ───────────────────────────────────────────────────
anthropic==0.86.0
openai==2.29.0
reportlab==4.4.10

# ── Ed25519 result signing ──────────────────────────────────────────────────
PyNaCl==1.6.2

# ── Transitive pins for CVE-fixed dependencies ──────────────────────────────
# pillow 12.2.0 fixes FITS decompression-bomb DoS (CVE-2026-40192).
pillow>=12.2.0
# python-multipart 0.0.26 fixes multipart preamble/epilogue DoS (CVE-2026-40347).
python-multipart>=0.0.26
# pygments 2.20.0 fixes ReDoS in AdlLexer (CVE-2026-4539).
pygments>=2.20.0

# ── Development & Testing ───────────────────────────────────────────────────
# pytest 9.0.3 fixes /tmp/pytest-of-{user} privilege issue (CVE-2025-71176).
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-cov==7.0.0
commitizen==4.13.9

# ── Dashboard ───────────────────────────────────────────────────────────────
fastapi==0.135.1
uvicorn==0.42.0

# ── Vast.ai burst compute ──────────────────────────────────────────────────
# vastai pins python-dateutil==2.6.1 which conflicts with pandas.
# Install separately: pip install --no-deps vastai
vastai==0.5.0

# ── Security Scanning (optional — used by security/security_scan.py) ────────
ruff==0.15.6
bandit==1.9.4
pip-audit>=2.7
detect-secrets==1.5.0
slowapi==0.1.9
